fix failure_test
