+++ test/failure_test.py
@@ -13,9 +13,6 @@ import unittest
 import ray.ray_constants as ray_constants
 import ray.test.test_functions as test_functions
 
-if sys.version_info >= (3, 0):
-    from importlib import reload
-
 
 def relevant_errors(error_type):
     return [info for info in ray.error_info() if info["type"] == error_type]
@@ -35,11 +32,23 @@ class TaskStatusTest(unittest.TestCase):
         ray.shutdown()
 
     def testFailedTask(self):
-        reload(test_functions)
+
+        @ray.remote
+        def throw_exception_fct1():
+            raise Exception("Test function 1 intentionally failed.")
+
+        @ray.remote
+        def throw_exception_fct2():
+            raise Exception("Test function 2 intentionally failed.")
+
+        @ray.remote(num_return_vals=3)
+        def throw_exception_fct3(x):
+            raise Exception("Test function 3 intentionally failed.")
+
         ray.init(num_workers=3, driver_mode=ray.SILENT_MODE)
 
-        test_functions.throw_exception_fct1.remote()
-        test_functions.throw_exception_fct1.remote()
+        throw_exception_fct1.remote()
+        throw_exception_fct1.remote()
         wait_for_errors(ray_constants.TASK_PUSH_ERROR, 2)
         self.assertEqual(
             len(relevant_errors(ray_constants.TASK_PUSH_ERROR)), 2)
@@ -47,7 +56,7 @@ class TaskStatusTest(unittest.TestCase):
             self.assertIn("Test function 1 intentionally failed.",
                           task.get("message"))
 
-        x = test_functions.throw_exception_fct2.remote()
+        x = throw_exception_fct2.remote()
         try:
             ray.get(x)
         except Exception as e:
@@ -56,7 +65,7 @@ class TaskStatusTest(unittest.TestCase):
             # ray.get should throw an exception.
             self.assertTrue(False)
 
-        x, y, z = test_functions.throw_exception_fct3.remote(1.0)
+        x, y, z = throw_exception_fct3.remote(1.0)
         for ref in [x, y, z]:
             try:
                 ray.get(ref)
